+++ packages/@valapi/valorant-api.com/src/service/Bundles.ts
@@ -2,13 +2,13 @@ import { ValorantApiComService } from "../client/ValorantApiComService";
 import type { ValorantApiCom } from "../client/ValorantApiCom";
 
 export namespace Bundles {
-    export interface Bundles {
+    export interface Bundles<L extends ValorantApiCom.Language> {
         uuid: string;
-        displayName: ValorantApiCom.Response.Languages<string>; // localized
-        displayNameSubText: ValorantApiCom.Response.Languages<string>; // localized
-        description: ValorantApiCom.Response.Languages<string>; // localized
-        extraDescription: ValorantApiCom.Response.Languages<string>; // localized
-        promoDescription: ValorantApiCom.Response.Languages<string>; // localized
+        displayName: ValorantApiComService.Languages<string, L>;
+        displayNameSubText: ValorantApiComService.Languages<string, L>;
+        description: ValorantApiComService.Languages<string, L>;
+        extraDescription: ValorantApiComService.Languages<string, L>;
+        promoDescription: ValorantApiComService.Languages<string, L>;
         useAdditionalContext: boolean;
         displayIcon: string;
         displayIcon2: string;
@@ -17,12 +17,12 @@ export namespace Bundles {
     }
 }
 
-export class Bundles extends ValorantApiComService {
-    public get(): Promise<ValorantApiCom.Response.Data<Bundles.Bundles[]>> {
+export class Bundles<L extends ValorantApiCom.Language = any> extends ValorantApiComService {
+    public get(): Promise<ValorantApiComService.Response<Bundles.Bundles<L>[]>> {
         return this.axios.get("/bundles");
     }
 
-    public getByUuid(uuid: string): Promise<ValorantApiCom.Response.Data<Bundles.Bundles>> {
+    public getByUuid(uuid: string): Promise<ValorantApiComService.Response<Bundles.Bundles<L>>> {
         return this.axios.get(`/bundles/${uuid}`);
     }
 }