+++ packages/@valapi/valorant-api.com/src/service/ContentTiers.ts
@@ -2,9 +2,9 @@ import { ValorantApiComService } from "../client/ValorantApiComService";
 import type { ValorantApiCom } from "../client/ValorantApiCom";
 
 export namespace ContentTiers {
-    export interface ContentTiers {
+    export interface ContentTiers<L extends ValorantApiCom.Language> {
         uuid: string;
-        displayName: ValorantApiCom.Response.Languages<string>; // localized
+        displayName: ValorantApiComService.Languages<string, L>;
         devName: string;
         rank: number;
         juiceValue: number;
@@ -15,12 +15,12 @@ export namespace ContentTiers {
     }
 }
 
-export class ContentTiers extends ValorantApiComService {
-    public get(): Promise<ValorantApiCom.Response.Data<ContentTiers.ContentTiers[]>> {
+export class ContentTiers<L extends ValorantApiCom.Language = any> extends ValorantApiComService {
+    public get(): Promise<ValorantApiComService.Response<ContentTiers.ContentTiers<L>[]>> {
         return this.axios.get("/contenttiers");
     }
 
-    public getByUuid(uuid: string): Promise<ValorantApiCom.Response.Data<ContentTiers.ContentTiers>> {
+    public getByUuid(uuid: string): Promise<ValorantApiComService.Response<ContentTiers.ContentTiers<L>>> {
         return this.axios.get(`/contenttiers/${uuid}`);
     }
 }