+++ packages/@valapi/valorant-api.com/src/service/Objectives.ts
@@ -1,5 +1,4 @@
 import { ValorantApiComService } from "../client/ValorantApiComService";
-import type { ValorantApiCom } from "../client/ValorantApiCom";
 
 export namespace Objectives {
     /**
@@ -13,11 +12,11 @@ export namespace Objectives {
 }
 
 export class Objectives extends ValorantApiComService {
-    public get(): Promise<ValorantApiCom.Response.Data<Objectives.Objectives[]>> {
+    public get(): Promise<ValorantApiComService.Response<Objectives.Objectives[]>> {
         return this.axios.get("/objectives");
     }
 
-    public getByUuid(uuid: string): Promise<ValorantApiCom.Response.Data<Objectives.Objectives>> {
+    public getByUuid(uuid: string): Promise<ValorantApiComService.Response<Objectives.Objectives>> {
         return this.axios.get(`/objectives/${uuid}`);
     }
 }