+++ packages/@valapi/valorant-api.com/src/service/Seasons.ts
@@ -2,9 +2,9 @@ import { ValorantApiComService } from "../client/ValorantApiComService";
 import type { ValorantApiCom } from "../client/ValorantApiCom";
 
 export namespace Seasons {
-    export interface Seasons {
+    export interface Seasons<L extends ValorantApiCom.Language> {
         uuid: string;
-        displayName: ValorantApiCom.Response.Languages<string>; // localized
+        displayName: ValorantApiComService.Languages<string, L>;
         type: string;
         startTime: string | Date;
         endTime: string | Date;
@@ -30,20 +30,20 @@ export namespace Seasons {
     }
 }
 
-export class Seasons extends ValorantApiComService {
-    public get(): Promise<ValorantApiCom.Response.Data<Seasons.Seasons[]>> {
+export class Seasons<L extends ValorantApiCom.Language = any> extends ValorantApiComService {
+    public get(): Promise<ValorantApiComService.Response<Seasons.Seasons<L>[]>> {
         return this.axios.get("/seasons");
     }
 
-    public getCompetitiveSeasons(): Promise<ValorantApiCom.Response.Data<Seasons.CompetitiveSeasons[]>> {
+    public getCompetitiveSeasons(): Promise<ValorantApiComService.Response<Seasons.CompetitiveSeasons[]>> {
         return this.axios.get("/seasons/competitive");
     }
 
-    public getByUuid(uuid: string): Promise<ValorantApiCom.Response.Data<Seasons.Seasons>> {
+    public getByUuid(uuid: string): Promise<ValorantApiComService.Response<Seasons.Seasons<L>>> {
         return this.axios.get(`/seasons/${uuid}`);
     }
 
-    public getCompetitiveSeasonByUuid(uuid: string): Promise<ValorantApiCom.Response.Data<Seasons.CompetitiveSeasons>> {
+    public getCompetitiveSeasonByUuid(uuid: string): Promise<ValorantApiComService.Response<Seasons.CompetitiveSeasons>> {
         return this.axios.get(`/seasons/competitive/${uuid}`);
     }
 }